requirements.txt
albumentations
efficientnet-pytorch==0.4.0
hyperopt
numpy
pandas
pillow
pyyaml
scikit-image
tqdm
torch==1.1.0
torchcontrib==0.0.2
torchvision==0.3.0
ujson
